+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/connector/client/IoTDBThriftSyncDeviceClientManager.java
@@ -1,0 +1,60 @@
+package org.apache.iotdb.db.pipe.connector.client;
+
+import org.apache.iotdb.common.rpc.thrift.TEndPoint;
+import org.apache.iotdb.commons.pipe.connector.client.IoTDBThriftSyncClientManager;
+import org.apache.iotdb.commons.pipe.connector.client.IoTDBThriftSyncConnectorClient;
+import org.apache.iotdb.tsfile.utils.Pair;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.List;
+
+public abstract class IoTDBThriftSyncDeviceClientManager extends IoTDBThriftSyncClientManager {
+
+  private static final Logger LOGGER =
+      LoggerFactory.getLogger(IoTDBThriftSyncDeviceClientManager.class);
+  private final LeaderCacheManager leaderCacheManager = new LeaderCacheManager();
+
+  protected IoTDBThriftSyncDeviceClientManager(
+      List<TEndPoint> endPoints,
+      boolean useSSL,
+      String trustStorePath,
+      String trustStorePwd,
+      boolean useLeaderCache) {
+    super(endPoints, useSSL, trustStorePath, trustStorePwd, useLeaderCache);
+  }
+
+  public Pair<IoTDBThriftSyncConnectorClient, Boolean> getClient(String deviceId) {
+    final TEndPoint endPoint = leaderCacheManager.getLeaderEndPoint(deviceId);
+    return useLeaderCache
+            && endPoint != null
+            && endPoint2ClientAndStatus.containsKey(endPoint)
+            && Boolean.TRUE.equals(endPoint2ClientAndStatus.get(endPoint).getRight())
+        ? endPoint2ClientAndStatus.get(endPoint)
+        : getClient();
+  }
+
+  public void updateLeaderCache(String deviceId, TEndPoint endPoint) {
+    if (!useLeaderCache) {
+      return;
+    }
+
+    try {
+      if (!endPoint2ClientAndStatus.containsKey(endPoint)) {
+        endPointList.add(endPoint);
+        endPoint2ClientAndStatus.put(endPoint, new Pair<>(null, false));
+        reconstructClient(endPoint);
+      }
+
+      leaderCacheManager.updateLeaderEndPoint(deviceId, endPoint);
+    } catch (Exception e) {
+      LOGGER.warn(
+          "Failed to update leader cache for device {} with endpoint {}:{}.",
+          deviceId,
+          endPoint.getIp(),
+          endPoint.getPort(),
+          e);
+    }
+  }
+}
