+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/connector/protocol/thrift/sync/IoTDBThriftSyncConnector.java
@@ -30,6 +30,7 @@ import org.apache.iotdb.commons.pipe.event.EnrichedEvent;
 import org.apache.iotdb.commons.utils.NodeUrlUtils;
 import org.apache.iotdb.db.conf.IoTDBConfig;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
+import org.apache.iotdb.db.pipe.connector.client.IoTDBThriftSyncDeviceClientManager;
 import org.apache.iotdb.db.pipe.connector.payload.evolvable.builder.IoTDBThriftSyncPipeTransferBatchReqBuilder;
 import org.apache.iotdb.db.pipe.connector.payload.evolvable.request.PipeTransferTabletBatchReq;
 import org.apache.iotdb.db.pipe.connector.payload.evolvable.request.PipeTransferTabletBinaryReq;
@@ -258,7 +259,9 @@ public class IoTDBThriftSyncConnector extends IoTDBSyncSslConnector {
       insertNode = pipeInsertNodeTabletInsertionEvent.getInsertNodeViaCacheIfPossible();
 
       if (insertNode != null) {
-        clientAndStatus = clientManager.getClient(insertNode.getDevicePath().getFullPath());
+        clientAndStatus =
+            ((IoTDBThriftSyncDeviceClientManager) clientManager)
+                .getClient(insertNode.getDevicePath().getFullPath());
         resp =
             clientAndStatus
                 .getLeft()
@@ -291,15 +294,16 @@ public class IoTDBThriftSyncConnector extends IoTDBSyncSslConnector {
               pipeInsertNodeTabletInsertionEvent, status));
     }
     if (insertNode != null && status.isSetRedirectNode()) {
-      clientManager.updateLeaderCache(
-          insertNode.getDevicePath().getFullPath(), status.getRedirectNode());
+      ((IoTDBThriftSyncDeviceClientManager) clientManager)
+          .updateLeaderCache(insertNode.getDevicePath().getFullPath(), status.getRedirectNode());
     }
   }
 
   private void doTransfer(PipeRawTabletInsertionEvent pipeRawTabletInsertionEvent)
       throws PipeException {
     final Pair<IoTDBThriftSyncConnectorClient, Boolean> clientAndStatus =
-        clientManager.getClient(pipeRawTabletInsertionEvent.getDeviceId());
+        ((IoTDBThriftSyncDeviceClientManager) clientManager)
+            .getClient(pipeRawTabletInsertionEvent.getDeviceId());
     final TPipeTransferResp resp;
 
     try {
@@ -327,8 +331,8 @@ public class IoTDBThriftSyncConnector extends IoTDBSyncSslConnector {
               pipeRawTabletInsertionEvent, status));
     }
     if (status.isSetRedirectNode()) {
-      clientManager.updateLeaderCache(
-          pipeRawTabletInsertionEvent.getDeviceId(), status.getRedirectNode());
+      ((IoTDBThriftSyncDeviceClientManager) clientManager)
+          .updateLeaderCache(pipeRawTabletInsertionEvent.getDeviceId(), status.getRedirectNode());
     }
   }
